wip: begin orbitdbplugin integration
