+++ src/components/OrbitDbHandler.vue
@@ -1,0 +1,4 @@
+<script>
+console.log("OrbitDbHandler");
+
+</script>
